Incrament

+++ LightFinder/Main.go
@@ -34,7 +34,7 @@ func robotRunLoop(gopigo3 *g.Driver, lightSensors [ 2 ]*aio.GroveLightSensorDriv
 			fmt.Errorf( "Error reading sensor1 %+v", error1 )
 		}
 		tolerence := 10
-		sensorDifference := math.Abs( sensor0Data - sensor1Data )
+		sensorDifference := int( math.Abs( float64( sensor0Data - sensor1Data ) ) )
 		if ( sensor0Data > sensor1Data && sensorDifference > tolerence ) || ( sensor0Data < 1000 && sensor1Data < 1000 ) {
 			CircleLeft( gopigo3, 10 )
 		} else if ( sensor0Data < sensor1Data && sensorDifference > tolerence ) {
